Split wasmer-package with two features: execution and authoring, for smaller builds
