+++ lib/package/src/lib.rs
@@ -1,7 +1,14 @@
 #[macro_use]
-#[cfg(test)]
+#[cfg(all(test, feature = "authoring"))]
 mod macros;
 
+mod error;
+
+pub use error::WasmerPackageError;
+
+#[cfg(feature = "authoring")]
 pub mod convert;
+#[cfg(feature = "authoring")]
 pub mod package;
+#[cfg(feature = "execution")]
 pub mod utils;
