+++ lib/swift/Cargo.toml
@@ -26,7 +26,7 @@ wasmer-wasix = { version = "=0.704.0", path = "../wasix", default-features = fal
   "sys",
 ] }
 webc.workspace = true
-wasmer-package = { workspace = true, features = ["webc-v1", "webc-v2", "webc-v3"] }
+wasmer-package = { workspace = true, features = ["execution", "webc-v1", "webc-v2", "webc-v3"] }
 
 
 [build-dependencies]
