+++ lib/wasix/Cargo.toml
@@ -20,7 +20,7 @@ harness = false
 required-features = ["sys-thread"]
 
 [dependencies]
-wasmer-package = { workspace = true, default-features = false, features = ["webc-v3"] }
+wasmer-package = { workspace = true, default-features = false, features = ["execution", "webc-v3"] }
 wasmer-wasix-types = { path = "../wasi-types", version = "0.704.0", features = [
   "enable-serde",
 ] }
@@ -218,6 +218,7 @@ ctrlc = ["tokio/signal"]
 # the minimal sys implementation
 sys-minimal = ["wasmer/sys", "sys-thread", "tokio/fs"]
 sys = [
+  "package-authoring",
   "webc/mmap",
   "time",
   "virtual-mio/sys",
@@ -242,6 +243,7 @@ sys = [
   "webc/v1",
   "webc/v2",
 ]
+package-authoring = ["wasmer-package/authoring"]
 sys-default = ["sys"]
 sys-poll = []
 extra-logging = []
